# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PlatinumCoin PIX Platform API
  version: 1.0.0
  description: |
    Public REST contract of the Pix platform. Conventions:
    - All endpoints (except /auth/login) require `Authorization: Bearer <JWT>`.
    - The **debited account is always derived from the JWT** (`accountId` claim); it is never accepted in any request body.
    - Money in request/response bodies is a decimal string in BRL (e.g. "125.50"); stored internally as integer cents.
    - `POST /payments/pix` is asynchronous: `202 Accepted` means "accepted for processing", not settled.
    - Errors follow RFC 7807 (application/problem+json) and **every** non-2xx carries `code` and
      `correlationId` — including the ones the framework produces before a handler runs (unknown
      route, wrong method, unsupported media type, unparseable body). See the `Problem` schema.
    - There is no API gateway in the local build: each path group is served by its own service
      (per-path `servers` overrides below) — auth :8081, pix-keys :8082, payments/accounts :8084,
      notifications :8087. The default server is payment-service.

    **Versioning (ARCHITECTURE §7.8).** URI versioning: every public route lives under `/v1`, enforced
    at build time by an ArchUnit rule in each service. Changes within `/v1` are **additive only** — a
    new optional request field, a new response field, a new endpoint, a new error `code` on an existing
...
    be migrated on our schedule.

...

...
servers:
...
paths:
...
  /accounts/me/statement/exports:
    post:
      tags: [accounts]
      summary: Request an asynchronous export of the COLD statement
      description: |
        The hot statement (`GET /accounts/me/statement`) serves the recent window straight out of
        the ledger. Everything older lives in the cold archive as monthly objects in object storage,
        and assembling a range of them is a slow read — seconds to minutes, not milliseconds. So it
        is modelled as a **request resource** rather than a blocking query: this call returns
        `202 Accepted` with an `exportId` and a `statusUrl`, a worker assembles the CSV, and the
        status endpoint flips from `PENDING` to `READY` with a time-limited `downloadUrl`
        (ARCHITECTURE §6.14).

        **Idempotency:** the `Idempotency-Key` header is REQUIRED, exactly as on `POST /payments/pix`
        and for the same reason — a client that retries a timed-out request must not start a second
        job. The same key with the same month range replays the original `202` and the original
        `exportId`. The same key with a **different** range returns `409 IDEMPOTENCY_KEY_REUSED`.
        Keys are scoped to the authenticated account.

        **Range rules**, all answered with `422`:
        - `fromMonth` must not be after `toMonth` (`INVALID_EXPORT_RANGE`)
        - at most 24 months (`INVALID_EXPORT_RANGE`)
        - no month before the account was opened (`INVALID_EXPORT_RANGE`)
        - a range lying **entirely inside the hot window** is refused with `USE_HOT_STATEMENT`
          rather than served: the data is already available synchronously, and building a second,
          slower copy of it would be duplicating the platform's own answer. A range that straddles
          the boundary is accepted and exports the cold part.

        Months in range for which the archive holds no object are **skipped, not failed** — an
        account simply had no movement that month. An export whose whole range is empty therefore
        succeeds with a header-only CSV.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fromMonth, toMonth]
              properties:
                fromMonth:
                  type: string
                  pattern: '^\d{4}-\d{2}$'
                  description: First month of the range, inclusive (`yyyy-MM`, UTC calendar months)
                  example: "2025-01"
                toMonth:
                  type: string
                  pattern: '^\d{4}-\d{2}$'
                  description: Last month of the range, inclusive
                  example: "2025-06"
      responses:
        '202':
          description: |
            Accepted for processing. Returned both on the first request and on every replay of the
            same key + range, with the same `exportId` — a replay is not distinguishable from the
            original by design.
          headers:
            Location:
              schema: { type: string }
              description: Same value as `statusUrl` in the body
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatementExportAccepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: IDEMPOTENCY_KEY_REUSED — the same key was replayed with a different month range
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422':
          description: |
            INVALID_EXPORT_RANGE — inverted, longer than 24 months, or starting before the account existed;
            USE_HOT_STATEMENT — the whole range is still inside the hot window, use `GET /accounts/me/statement`
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /statement-exports/{exportId}:
    get:
      tags: [accounts]
      summary: Poll an export, and download it once it is READY
      description: |
        The polling half of the `202` contract. Ownership is enforced from the JWT: an export that
        belongs to another account answers `404`, never `403` — the existence of another customer's
        export is not a fact this API discloses.

        While `PENDING` the response carries `Retry-After: 5`, which is the platform telling a
        front-end how to poll politely rather than leaving it to guess.
      parameters:
        - name: exportId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Current state of the export
          headers:
            Retry-After:
              schema: { type: integer, example: 5 }
              description: Present on `PENDING` only
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatementExport' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A **user** token from `POST /v1/auth/login` — a JWT carrying `sub` (userId), `accountId` and
        `typ: user` (ADR-0007, amended by ADR-0017).

        Every operation in this contract is customer-facing and accepts **only** `typ=user`. The
        platform's other token type, `typ=service`, is minted by a service for one internal call and
        is refused here with `403 PUBLIC_ROUTE_FORBIDDEN` — the two surfaces are deliberately disjoint
        in both directions, so a leaked service credential cannot be replayed against the customer API.

        The `/internal/**` service-to-service ports are **not part of this contract** (no client calls
        one). They accept `typ=service` only, validating `aud` and an operation `scope`, and answer
        `403 INTERNAL_PORT_FORBIDDEN` to anything else. See `SECURITY.md` and ADR-0017.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }
      description: Client-generated UUID, one per business operation; reuse it on retries.
  schemas:
...
    StatementExportAccepted:
      type: object
      properties:
        exportId:
          type: string
          description: >-
            Server-minted id of the export request. Derived deterministically from the account and
            the `Idempotency-Key`, which is what makes a replay return the same id without a second
            store to consult; treat it as opaque.
          example: exp-4c2a9f61b0d84e7a9c13f8ab27d5e604
        status:
          type: string
          enum: [PENDING, READY, FAILED]
          example: PENDING
          description: >-
            `PENDING` on a first request. On a replay this reports the export's **current** state, which
            may already be `READY` — telling a client `PENDING` about a downloadable export would be a
            lie it would then poll to discover.
        statusUrl:
          type: string
          description: >-
            Path of `GET /statement-exports/{exportId}`, relative to this service — the same value as
            the `Location` header. A path and not an absolute URL because the service knows its own
            routes but not the public base URL it is reached on (the same choice `POST /payments/pix`
            makes for its `Location`).
          example: /v1/statement-exports/exp-4c2a9f61b0d84e7a9c13f8ab27d5e604
    StatementExport:
      type: object
      properties:
        exportId: { type: string }
        status:
          type: string
          enum: [PENDING, READY, FAILED]
          description: >-
            `PENDING` while the worker assembles the artifact, then exactly one of `READY` or
            `FAILED`. Both are terminal — the transition is guarded, so a redelivered queue message
            cannot move a finished export or produce a second artifact.
        requestedRange:
          type: object
          properties:
            fromMonth: { type: string, example: "2025-01" }
            toMonth: { type: string, example: "2025-06" }
        requestedAt: { type: string, format: date-time }
        downloadUrl:
          type: string
          nullable: true
          description: >-
            Presigned URL of the CSV artifact. Present on `READY` only, and **minted fresh for each
            response** with a 1h lifetime from *this read*. The platform stores the object key, never a
            URL: a link signed when the worker finished would start expiring while the customer was
            still being told the file was ready, and an export whose only handle had expired would be
            permanently undownloadable. A client that finds a link stale simply polls again.
        expiresAt:
          type: string
          format: date-time
          nullable: true
          description: When `downloadUrl` stops working. Present on `READY` only.
        failureReason:
          type: string
          nullable: true
          description: Present on `FAILED` only — why the assembly could not be completed.
...
    Problem:
      description: |
        RFC 7807 problem detail. The platform adds two members to every error it returns, without
        exception — `scripts/error-contract-audit.sh` walks the running stack and asserts it, and
        `ErrorContractIT` asserts it in-process.

        `code` is the machine contract: branch on it, never on `detail`, which is English prose written
        for a human and may be reworded at any time. `correlationId` is the id echoed on the
        `X-Correlation-Id` response header; quoting it in a support ticket is what makes one `grep`
        reconstruct the request across every service (ADR-0012).

        New `code` values may appear within `/v1` — that is an additive change (ARCHITECTURE §7.8), so
        treat an unrecognised code as its HTTP status class rather than as a parse failure.
      type: object
      required: [type, title, status, code, correlationId]
      properties:
        type: { type: string, example: 'about:blank' }
        title: { type: string, example: Unprocessable Entity }
        status: { type: integer, example: 422 }
        detail:
          description: Human-readable explanation. Not a contract — do not parse or branch on it.
          type: string
        code:
          description: |
            Stable, machine-readable reason. Domain codes come from the service that owns the rule
            (`LIMIT_EXCEEDED`, `INSUFFICIENT_FUNDS`, `FRAUD_DENIED`, `IDEMPOTENCY_KEY_REUSED`,
            `KEY_NOT_FOUND`, `PAYMENT_NOT_FOUND`, `DIRECTORY_UNAVAILABLE`, `LEDGER_UNAVAILABLE`, …).
            Rejections that happen before any handler runs are derived from the status
            (`NOT_FOUND`, `METHOD_NOT_ALLOWED`, `UNSUPPORTED_MEDIA_TYPE`, …), with two named
            exceptions on 400, where one status carries two meanings a client fixes differently:
            `MALFORMED_REQUEST` (the body did not parse) and `VALIDATION_ERROR` (it parsed and its
            fields, query parameters or required headers were wrong).
          type: string
          example: LIMIT_EXCEEDED
        correlationId:
          description: Same value as the `X-Correlation-Id` response header.
          type: string
          format: uuid
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing/invalid JWT
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
security:
  - bearerAuth: []